+++ src/inpage/QryptoRPCProvider.ts
@@ -1,6 +1,6 @@
 import { rpcCall } from './rpcCall';
 
-export class QryptoRPCProvider {
+export class QryptoRpcProvider {
   public rawCall(method: string, args: any[]) {
     return rpcCall(method, args);
   }
