+++ src/inpage/rpcCall.ts
@@ -1,8 +1,6 @@
 import { API_TYPE } from '../constants';
-import { requestExtensionAPI } from './utils';
-import { IRPCCallResponsePayload, IRPCCallRequestPayload } from '../types';
-
-const processingRPCCallRequests: { [id: string]: IRPCCallRequest } = {};
+import { processingRPCCallRequests, requestExtensionAPI } from './utils';
+import { IRPCCallRequestPayload } from '../types';
 
 export function rpcCall(method: string, args: any[]): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -14,26 +12,4 @@ export function rpcCall(method: string, args: any[]): Promise<string> {
       payload: { method, args, id },
     });
   });
-}
-
-export function handleRPCCallResponse(response: IRPCCallResponsePayload) {
-  const request = processingRPCCallRequests[response.id];
-  if (!request) {
-    return;
-  }
-
-  delete processingRPCCallRequests[response.id];
-
-  if (response.error != null) {
-    request.reject(response.error);
-
-    return;
-  }
-  console.log('call rpc method success!');
-  request.resolve(response.result);
-}
-
-interface IRPCCallRequest {
-  resolve: (result?: any) => void;
-  reject: (reason?: any) => void;
 }
